services/mqtt-bridge/requirements.txt
# ForgeLink MQTT Bridge Dependencies

# MQTT
paho-mqtt==2.1.0

# Kafka
confluent-kafka==2.4.0

# Configuration
pydantic==2.7.4
pydantic-settings==2.3.3
python-dotenv==1.0.1

# Observability
prometheus-client==0.20.0
structlog==24.2.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
